fix(builder): ssh timeout to prevent intermediary firewalls/loadbalancers from dropping connection
Currently the Deis-builder loses connection to cilents sometimes when tbere
are intermediaries that monitors the connection and drops it after a certain
interval (eg AWS Load Balancers have a 5 min timeout) while the Builder is
busy with some activities such as tarring a file etc.

This will ensure a ssh-keepalive is sent every minute to keep the connection
open.

TODO: Tests needs to be created with mock Firewalls in the middle that are
configured to drop connection as specific intervals, and we need to demonstrate
that the builder does not lose connection past those period when connected
to by a client.

closes #1046, #1195