fix(builder): ssh timeout to prevent intermediary firewalls/loadbalancers from dropping connection
Currently the Deis-builder loses connection to cilents sometimes when tbere
are intermediaries that monitors the connection and drops it after a certain
interval (eg AWS Load Balancers have 1 min timeouts) while the Builder is
busy with some activities such as tarring a file etc.

This will ensure a ssh-keepalive is sent every minute to keep the connection
open.

TESTING: The added go-spec needs to be filled in to do a real automated test.
Otherwise similar steps can be followed to do a manual test.

closes #1046, #1195